libdevmapper: add statistics data structures and interface
Add data structures, type definitions and interfaces to
libdevmapper to work with device-mapper statistics.

To simplify handling gaps in the sequence of region_ids for users
of the library, and to allow for future changes in the data
structures used to contain statistics values in userspace, the data
structures themselves are not exported in libdevmapper.h.

Instead an opaque handle of type struct dm_stats* is obtained by
calling the library and all subsequent statistics operations are
carried out using this handle. A dm_stats object represents the
complete set of available counter sets for an individual mapped
device.

The dm_stats handle contains a pointer to a table of one or more
dm_stats_region objects representing the regions registered with the
@stats_create message. These in turn point to a table of one or more
dm_stats_counters objects containing the counter sets for each defined
area within the region:

  dm_stats->dm_stats_region[nr_regions]->dm_stats_counters[nr_areas]

This structure is private to the library and may change in future
versions: all users should make use of the public interface and treat
the dm_stats type as an opaque handle. Accessor methods are provided
to obtain values stored in individual region and area objects.

Ranges and counter sets are stored in order of increasing device
sector.

Public methods are provided to create and destroy handles and to
list, create, and destroy, statistics regions as well as to obtain and
parse the counter data.

Linux iostat-style derived performance metrics are provided to return
higher-level performance metrics:

    dm_stats_get_throughput()
    dm_stats_get_utilization()
    dm_stats_get_service_time()
    dm_stats_get_rd_merges_per_sec()
    dm_stats_get_wr_merges_per_sec()
    dm_stats_get_reads_per_sec()
    dm_stats_get_read_sectors_per_sec()
    dm_stats_get_writes_per_sec()
    dm_stats_get_write_sectors_per_sec()
    dm_stats_get_average_request_size()
    dm_stats_get_average_queue_size()
    dm_stats_get_await()
    dm_stats_get_r_await()
    dm_stats_get_w_await()

+++ libdm/libdm-report.c
@@ -4271,19 +4271,19 @@ void dm_report_set_interval(struct dm_report *rh, uint64_t interval)
 	rh->interval = interval;
 }
 
-void dm_report_set_interval_ms(struct dm_report *rh, uint32_t interval_ms)
+void dm_report_set_interval_ms(struct dm_report *rh, uint64_t interval_ms)
 {
 	rh->interval = interval_ms * NSEC_PER_MSEC;
 }
 
-uint32_t dm_report_get_interval(struct dm_report *rh)
+uint64_t dm_report_get_interval(struct dm_report *rh)
 {
 	return rh->interval;
 }
 
-uint32_t dm_report_get_interval_ms(struct dm_report *rh)
+uint64_t dm_report_get_interval_ms(struct dm_report *rh)
 {
-	return (uint32_t) (rh->interval / NSEC_PER_MSEC);
+	return (rh->interval / NSEC_PER_MSEC);
 }
 
 uint64_t dm_report_get_last_interval(struct dm_report *rh)
